backend: rename SimulationSocketListener -> SimulationNamespaceListener
